requirements.txt
alembic==1.9.1
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==3.4.8
fastapi==0.89.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
protobuf==4.21.12
pycparser==2.21
pydantic==1.10.4
PyJWT==1.7.1
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
